Implement LaTeX file parser for citations
Find all citations (possibly following nested inputs) in a given LaTeX
file.  This is implemented using the classical lexer/token logic, as I
intuitively found it to be more apt for this than regular expressions.
However, I'm not sure what the best way to do it is.  All I know is
that it does not need to be a full parser because we only care about a
couple of different commands.