Refactor Citations struct and computation of gathering citations
Previously, we had a Citations struct that essentially wrapped a
HashSet, which just contained the keys we found from either LaTeX or
Bibliography source, and that was computed twice from source (as
required).  This was okay for the --unused functionality, as that only
required us to know the citation keys and nothing else.  However, we
are now required to know some information from the Bibliography file.

The new Citations struct wraps a HashMap<String, Citation>, and
Citation contains (importantly) a key, and (optionally) an entry (from
the Bibliography file), and (optionally) a citation command.  If the
citation *command* is present then we have obtained the citation from
LaTeX source; but if the entry is present, then we have obtained it
from the Bibliography source.

The previous version of the Citations struct was originally
implemented in 59359ea.  This is one modification to it; however, I
will need to modify it again as some functions might not be required
to read LaTeX source at all, so we shouldn't do this if we don't have
to (especially considering it might not even be relevant in some
cases, if the user doesn't have any LaTeX source document).

Another consideration is to have something like
```rust
struct Citation {
    pub key: String,
    entry: Option<Entry>,
    uses: Vec<LaTeXCitation>,
}
```
And perhaps give LaTeXCitation from latex.rs some more information
```rust
struct Loc {
    line: usize,
    cite_cmd_col: usize,
    cite_key_col: usize,
}

struct LaTeXCitation {
    key: String,
    pub cite_cmd: String,
    loc: Loc,
}
```

But that would still require us to parse both LaTeX and Bibliography
source, when it may not be required by the functionality of the
tool (depending no how it's being used).  We'll see how the
requirements of the project evolve as we keep developing it.